refactor(security): split mod.rs into bus/alert/auth_event submodules
Cycle 33 LOC pass: src/security/mod.rs was 335 LOC. Split into:
- bus.rs (broadcast channel + env mode flags)
- alert.rs (SecurityAlert model + builder)
- auth_event.rs (AuthEvent + log_auth_event)

Public API preserved via re-exports.
